bluetooth: Add support for the second version acaia scale protocol
The acaia scales (in protocol version 2) provide a BLE Internet Protocol Support
Serice which includes an "age" characteristic. By enabeling notifications on this
characteristic we can read messages from the scale.
The scale needs to be configured to send weight, timer and battery messages if
we want them. We also need to send an identifier to the scale which supposetly
is meant to be used as an unique identifier of the calling app. We are using a
dummy value.
Furthermore if no transmission happened on the channel we need to send an heartbeat
package ensuring the scale our aliveness and keeping the connection in tact.

All Acaia specifics are encapsuled into their respective functions, the main read
loop is integrated into the atomax and decent scale functionality.

This was tested with an acaia lunar with unknown firmware version and the original
decent tablet with the most recent software.

Signed-off-by: Mimoja <git@mimoja.de>